requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dnspython==1.16.0
entrypoints==0.3
eventlet==0.24.1
ffmpeg==1.4
Flask==1.0.2
Flask-SocketIO==3.3.2
greenlet==0.4.15
h5py==2.9.0
html5lib==0.9999999
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
monotonic==1.5
moviepy==1.0.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.16.3
opencv-contrib-python==4.1.0.25
opencv-contrib-python-headless==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
proglog==0.1.9
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-engineio==3.5.1
python-socketio==3.1.2
pytz==2019.1
PyWavelets==1.0.3
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
